chore: update OpenTelemetry Python requirements
Update the OpenTelemetry Python requirements for this project to support
external instrumentation using opentelemetry-instrumentation.

This allows users to specify their own tracers and inject them into this
application if needed. Otherwise, a no-op tracer is used, which will not
collect any traces.

This commit also updates the requirements.txt file to explicitly include
version 1.14 of different OpenTelemetry libraries. Note that we are not
using the latest 1.15 release due to an issue with certain imports (see
https://bit.ly/3Wz9DXJ for more details).