chore: move PyPI dev installs to post-create.sh
Move all PyPI installations for the devcontainer from the Dockerfile to
the post-create.sh script. Python and pip are installed as devcontainer
features, which are configured after the Dockerfile is executed. If
Python is not already available on the image used, pip commands in the
Dockerfile will fail.

By moving the pip command invocation to the post-create.sh script, we
ensure that the Python feature is available, regardless of the container
image used.